requirements.txt
typer[all]
pyyaml